#268 Optimize the `autoWidth` calculation method to make the column width more accurate

+++ src/main/java/org/ttzero/excel/entity/Column.java
@@ -97,7 +97,7 @@ public class Column {
      * The cell width
      */
     public double width;
-    public int o;
+    public double o;
     public Styles styles;
     public Comment headerComment, cellComment;
     /**
